Extract middle sources to separate folder
