+++ middle/middle-ram.hpp
